feat: implementa AtendimentoRemotoService

+++ src/main/java/com/gestaooserp/dev/repository/AtendimentoRemotoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.AtendimentoRemoto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtendimentoRemotoRepository extends JpaRepository<AtendimentoRemoto,Integer> {
+public interface AtendimentoRemotoRepository extends JpaRepository<AtendimentoRemoto,Long> {
 }
